+++ nitrocli/src/args.rs
@@ -30,10 +30,10 @@ type Result<T> = result::Result<T, Error>;
 /// A top-level command for nitrocli.
 #[derive(Debug)]
 pub enum Command {
-  Clear,
   Config,
   Lock,
   Otp,
+  Pin,
   Pws,
   Status,
   Storage,
@@ -43,10 +43,10 @@ impl Command {
   /// Execute this command with the given arguments.
   pub fn execute(&self, args: Vec<String>) -> Result<()> {
     match *self {
-      Command::Clear => clear(args),
       Command::Config => config(args),
       Command::Lock => lock(args),
       Command::Otp => otp(args),
+      Command::Pin => pin(args),
       Command::Pws => pws(args),
       Command::Status => status(args),
       Command::Storage => storage(args),
@@ -60,10 +60,10 @@ impl fmt::Display for Command {
       f,
       "{}",
       match *self {
-        Command::Clear => "clear",
         Command::Config => "config",
         Command::Lock => "lock",
         Command::Otp => "otp",
+        Command::Pin => "pin",
         Command::Pws => "pws",
         Command::Status => "status",
         Command::Storage => "storage",
@@ -77,10 +77,10 @@ impl str::FromStr for Command {
 
   fn from_str(s: &str) -> result::Result<Self, Self::Err> {
     match s {
-      "clear" => Ok(Command::Clear),
       "config" => Ok(Command::Config),
       "lock" => Ok(Command::Lock),
       "otp" => Ok(Command::Otp),
+      "pin" => Ok(Command::Pin),
       "pws" => Ok(Command::Pws),
       "status" => Ok(Command::Status),
       "storage" => Ok(Command::Storage),
@@ -284,6 +284,42 @@ impl From<OtpMode> for nitrokey::OtpMode {
 }
 
 #[derive(Debug)]
+enum PinCommand {
+  Clear,
+}
+
+impl PinCommand {
+  fn execute(&self, args: Vec<String>) -> Result<()> {
+    match *self {
+      PinCommand::Clear => pin_clear(args),
+    }
+  }
+}
+
+impl fmt::Display for PinCommand {
+  fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+    write!(
+      f,
+      "{}",
+      match *self {
+        PinCommand::Clear => "clear",
+      }
+    )
+  }
+}
+
+impl str::FromStr for PinCommand {
+  type Err = ();
+
+  fn from_str(s: &str) -> result::Result<Self, Self::Err> {
+    match s {
+      "clear" => Ok(PinCommand::Clear),
+      _ => Err(()),
+    }
+  }
+}
+
+#[derive(Debug)]
 enum PwsCommand {
   Clear,
   Get,
@@ -441,15 +477,6 @@ fn storage_status(args: Vec<String>) -> Result<()> {
   parse(&parser, args)?;
 
   commands::storage_status()
-}
-
-/// Clear the PIN as cached by various other commands.
-fn clear(args: Vec<String>) -> Result<()> {
-  let mut parser = argparse::ArgumentParser::new();
-  parser.set_description("Clears the cached passphrases");
-  parse(&parser, args)?;
-
-  commands::clear()
 }
 
 /// Execute a config subcommand.
@@ -718,6 +745,39 @@ fn otp_status(args: Vec<String>) -> Result<()> {
   commands::otp_status(all)
 }
 
+/// Execute a PIN subcommand.
+fn pin(args: Vec<String>) -> Result<()> {
+  let mut subcommand = PinCommand::Clear;
+  let mut subargs = vec![];
+  let mut parser = argparse::ArgumentParser::new();
+  parser.set_description("Manages the Nitrokey PINs");
+  let _ = parser.refer(&mut subcommand).required().add_argument(
+    "subcommand",
+    argparse::Store,
+    "The subcommand to execute (clear)",
+  );
+  let _ = parser.refer(&mut subargs).add_argument(
+    "arguments",
+    argparse::List,
+    "The arguments for the subcommand",
+  );
+  parser.stop_on_first_argument(true);
+  parse(&parser, args)?;
+  drop(parser);
+
+  subargs.insert(0, format!("nitrocli pin {}", subcommand));
+  subcommand.execute(subargs)
+}
+
+/// Clear the PIN as cached by various other commands.
+fn pin_clear(args: Vec<String>) -> Result<()> {
+  let mut parser = argparse::ArgumentParser::new();
+  parser.set_description("Clears the cached PINs");
+  parse(&parser, args)?;
+
+  commands::pin_clear()
+}
+
 /// Execute a PWS subcommand.
 fn pws(args: Vec<String>) -> Result<()> {
   let mut subcommand = PwsCommand::Get;
@@ -858,7 +918,7 @@ fn parse_arguments(args: Vec<String>) -> Result<(Command, Vec<String>)> {
   let _ = parser.refer(&mut command).required().add_argument(
     "command",
     argparse::Store,
-    "The command to execute (clear|config|lock|otp|status|storage)",
+    "The command to execute (config|lock|otp|pin|pws|status|storage)",
   );
   let _ = parser.refer(&mut subargs).add_argument(
     "arguments",